fix(controller/client): remove baseURL variable
This variable stored an older version of the controller URL before
appending the path, which confused users which endpoint the client
was hitting when testing a v2 client against a v1 cluster.

Old error message:

```
http://deis.10.245.1.3.xip.io does not appear to be a valid Deis controller. Make sure that the
Controller URI is correct and the server is running.
```

New error message:

```
http://deis.10.245.1.3.xip.io/v2/ does not appear to be a valid Deis controller. Make sure that the
Controller URI is correct and the server is running.
```

Note the /v2/ path appended to the URL, which users can assume this is a v2 client.